refactor(profile): Move Geo pagination to 'common'
- Relocate 'pagination' from shared 'app' to 'geo/common'
