+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/city/repository/CustomCityRepository.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.city.repository;
 
-import ru.pavbatol.myplace.app.pagination.GeoEntityPagingRepository;
+import ru.pavbatol.myplace.geo.common.pagination.GeoEntityPagingRepository;
 import ru.pavbatol.myplace.geo.city.model.City;
 
 public interface CustomCityRepository extends GeoEntityPagingRepository<City> {
