+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/city/repository/CustomCityRepositoryImpl.java
@@ -1,7 +1,7 @@
 package ru.pavbatol.myplace.geo.city.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import ru.pavbatol.myplace.app.pagination.AbstractGeoEntityPagingRepository;
+import ru.pavbatol.myplace.geo.common.pagination.AbstractGeoEntityPagingRepository;
 import ru.pavbatol.myplace.geo.city.model.City;
 
 import javax.persistence.EntityManager;
