+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/common/pagination/Sliced.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.app.pagination;
+package ru.pavbatol.myplace.geo.common.pagination;
 
 import lombok.AccessLevel;
 import lombok.Getter;
