+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/district/repository/CustomDistrictRepository.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.district.repository;
 
-import ru.pavbatol.myplace.app.pagination.GeoEntityPagingRepository;
+import ru.pavbatol.myplace.geo.common.pagination.GeoEntityPagingRepository;
 import ru.pavbatol.myplace.geo.district.model.District;
 
 public interface CustomDistrictRepository extends GeoEntityPagingRepository<District> {
