+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/house/repository/CustomHouseRepository.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.house.repository;
 
-import ru.pavbatol.myplace.app.pagination.GeoEntityPagingRepository;
+import ru.pavbatol.myplace.geo.common.pagination.GeoEntityPagingRepository;
 import ru.pavbatol.myplace.geo.house.model.House;
 
 public interface CustomHouseRepository extends GeoEntityPagingRepository<House> {
