+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/street/repository/CustomStreetRepository.java
@@ -1,6 +1,6 @@
 package ru.pavbatol.myplace.geo.street.repository;
 
-import ru.pavbatol.myplace.app.pagination.GeoEntityPagingRepository;
+import ru.pavbatol.myplace.geo.common.pagination.GeoEntityPagingRepository;
 import ru.pavbatol.myplace.geo.street.model.Street;
 
 public interface CustomStreetRepository extends GeoEntityPagingRepository<Street> {
